Create README.md
# Leadtracker-Chrome-extension
An amazing lead tracker chrome extension. Which allows users to keep track of any given URL and help save them in the given user local storage (chrome database). Additionally, It was built with HTML, Javascript and CSS. Furthermore, it gives user an option to save tab, to input and save any given URL. Lastly, the Lead tracker extension also have the option of deleting the save URL on your local storage when a user double click 
on the delete button.